Refactor homeworks app
