+++ src/apps/homeworks/api/serializers/__init__.py
@@ -1,0 +1,1 @@
+from .homework_serializers import HomeworkViewSerializer, HomeworkCreationSerializer
