+++ src/apps/homeworks/api/urls.py
@@ -1,0 +1,11 @@
+from rest_framework.routers import DefaultRouter
+
+from apps.homeworks.api.views import HomeworkViewSet
+
+router = DefaultRouter()
+
+router.register(r"homeworks", HomeworkViewSet, basename="homeworks")
+
+urlpatterns = []
+
+urlpatterns += router.urls
