+++ src/apps/homeworks/api/views/__init__.py
@@ -1,0 +1,1 @@
+from .homework_view_set import HomeworkViewSet
