+++ src/apps/homeworks/models/__init__.py
@@ -1,0 +1,1 @@
+from .homework import Homework
